refac: 추가정보 수정 기능 도메인 변경
